Revert "[FEATURE] 게시글 작성 모달 및 게시글 작성 기능 구현"
