+++ next.config.mjs
@@ -1,8 +1,4 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {
-  images: {
-    domains: ["saypjauxxtdemjrnqqtg.supabase.co"], // 허용할 외부 도메인 추가
-  },
-};
+const nextConfig = {};
 
 export default nextConfig;
