+++ package.json
@@ -6,8 +6,7 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "generate-types": "npx supabase gen types typescript --project-id saypjauxxtdemjrnqqtg --schema public > types_db.ts"
+    "lint": "next lint"
   },
   "dependencies": {
     "@heroicons/react": "^2.1.5",
