+++ src/app/(authorized)/@modal/default.tsx
@@ -1,3 +1,3 @@
-export default function Default() {
-  return null;
+export default function Default() { 
+    return null
 }
