--- src/components/button/index.ts
@@ -1,2 +1,0 @@
-export { default as FileUploadButton } from "./FileUploadButton";
-export { default as SubmitButton } from "./SubmitButton";
