+++ src/components/modal.tsx
@@ -1,0 +1,42 @@
+"use client";
+
+import { XMarkIcon } from "@heroicons/react/24/solid";
+import { useRouter } from "next/navigation";
+import { useEffect, useRef } from "react";
+import { createPortal } from "react-dom";
+
+export default function Modal({ children }: { children: React.ReactNode }) {
+  const dialogRef = useRef<HTMLDialogElement>(null);
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!dialogRef.current?.open) {
+      dialogRef.current?.showModal();
+      dialogRef.current?.scrollTo({ top: 0 });
+    }
+  }, []);
+
+  return createPortal(
+    <div
+      className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50 z-50"
+      onClick={(e) => {
+        if (e.target === e.currentTarget) {
+          router.back();
+        }
+      }}
+    >
+      <div className="relative bg-white p-6 rounded-lg w-full max-w-lg">
+        <button
+          className="absolute right-6"
+          onClick={() => {
+            router.back();
+          }}
+        >
+          <XMarkIcon className="w-6 h-6 text-gray-500" />
+        </button>
+        {children}
+      </div>
+    </div>,
+    document.getElementById("modal-root") as HTMLElement
+  );
+}
