--- src/components/modal/Modal.module.css
@@ -1,11 +1,0 @@
-.modalContainer {
-  @apply relative;
-}
-
-.modalOverlay {
-  @apply absolute inset-0 flex items-center justify-center bg-black bg-opacity-50 z-50;
-}
-
-.modalContent {
-  @apply relative bg-white p-6 rounded-lg w-full max-w-xl max-h-fit;
-}
