--- src/lib/feed/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./useCreateFeed";
