+++ src/stores/auth/useAuthStore.ts
@@ -18,10 +18,11 @@ export const useAuthStore = create<AuthStore>((set) => ({
           if (event === "SIGNED_IN" && session?.user) {
             set({
               user: {
-                id: session.user.id,
-                email: session.user.email || "",
-                nickname: session.user.user_metadata?.nickname || "",
-                profile_image: session.user.user_metadata?.profile_image || "",
+                id: session?.user?.id,
+                email: session?.user?.email || "",
+                nickname: session?.user?.user_metadata?.nickname || "",
+                profile_image:
+                  session?.user?.user_metadata?.profile_image || null,
               },
               isLogin: true,
             });
