--- src/utils/supabase/storage.ts
@@ -1,3 +1,0 @@
-export function getImageUrl(path: string) {
-  return `${process.env.NEXT_PUBLIC_SUPABASE_URL}/storage/v1/object/public/${process.env.NEXT_PUBLIC_STORAGE_BUCKET}/${path}`;
-}
